Reorganize Orphans module to avoid cyclical dependencies
